Crate hl_capture

Modules

capture

command

cvar

dl

encode

engine

fps_converter

hooks

macros

sdl

utils

Functions

CL_Disconnect⚠
Stops the currently running game, returning to the main menu.

Con_ToggleConsole_f⚠
Handler for the toggleconsole command.

GL_SetMode⚠
Sets up the main FBOs and the display mode.

Host_FilterTime⚠
Calculates the frame time and limits the FPS.

Key_Event⚠
Handles key callbacks.

Memory_Init⚠
Initializes the hunk memory.

RunListenServer⚠
The "main" function of hw.so, called inside CEngineAPI::Run().

S_PaintChannels⚠
Mixes sound into the output buffer using the paintbuffer.

S_TransferStereo16⚠
Transfers the contents of the paintbuffer into the output buffer.

Sys_VID_FlipScreen⚠
Flips the screen.

VideoMode_IsWindowed⚠
Returns whether the game is running in windowed mode.