Module hl_capture::engine

Structs

Engine
A "container" for unsafe engine functions.

EngineCVarGuard
A guard for statically ensuring that no engine functions are called while the engine CVar reference is valid. Holds a mutable reference.

MainThreadData
Global variables accessible from the main game thread.

MainThreadDataContainer
A Send+Sync container to allow putting MainThreadData into a global variable.

MainThreadMarker
A static guarantee of being in the main game thread, separate from Engine to allow borrowing.

Statics

MAIN_THREAD_DATA

Type Definitions

Result