Struct hl_capture::engine::Engine

pub struct Engine {
    _private: PhantomData<*const ()>,
}

A "container" for unsafe engine functions.

It's used for exposing safe interfaces for these functions where they can be used safely (for example, in console command handlers).

Fields
_private: PhantomData<*const ()>

This field serves two purposes: firstly, it prevents creating the struct not via the unsafe new() method, and secondly, it marks the struct as !Send and !Sync.

Methods

impl Engine

unsafe fn new() -> Self

Creates an instance of Engine.

Safety

Unsafe because this function should only be called from the main game thread.

fn args(&self) -> Args

Returns an iterator over the console command arguments.

fn con_print(&self, string: &str)

Prints the given string to the game console.

fn cmd_argc(&self) -> u32

Returns the number of console command arguments.

fn cmd_argv(&self, index: u32) -> String

Returns the console command argument with the given index.