Struct hl_capture::fps_converter::sampling::SamplingConverterPrivate

struct SamplingConverterPrivate {
    ocl_runtime_data: MaybeUnavailable<OclRuntimeData>,
    ocl_backup_buffer: Option<Vec<Float>>,
    video_resolution: (u32, u32),
    gl_sampling_buffer: Vec<f32>,
    gl_read_buffer: Vec<u8>,
}

Data with a destructor.

Fields
ocl_runtime_data: MaybeUnavailable<OclRuntimeData>

Data used for OpenCL operations.

This is Some(None) if OpenCL is unavailable and None during an engine restart.

ocl_backup_buffer: Option<Vec<Float>>

Pixels from the buffer are stored here when the engine restarts.

video_resolution: (u32, u32)

The video resolution.

gl_sampling_buffer: Vec<f32>

The OpenGL sampling buffer.

gl_read_buffer: Vec<u8>

The OpenGL read buffer.

Methods

impl SamplingConverterPrivate

fn new(engine: &mut Engine, video_resolution: (u32, u32)) -> Self

fn get_ocl_data(&mut self, engine: &mut Engine) -> Option<&mut OclRuntimeData>

fn backup_and_free_ocl_data(&mut self, engine: &mut Engine)

This should be called before an engine restart.

fn restore_ocl_data(&mut self, engine: &mut Engine)

This should be called after an engine restart.