Struct hl_capture::function::Function

pub struct Function<F> {
    ptr: F,
}

A container for a FFI function pointer.

Function<F> ensures the stored pointer is valid at all times. By default it is set to a special function which panics upon being called. This, unfortunately, results in varargs functions not being storable.

Fields
ptr: F

The stored function pointer.

Methods

impl<R, A, B, C, D, E, F> Function<unsafe fn(_: A, _: B, _: C, _: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, a: A, b: B, c: C, d: D, e: E, f: F) -> R

fn get(&self) -> unsafe fn(_: A, _: B, _: C, _: D, _: E, _: F) -> R

fn set(&mut self, f: unsafe fn(_: A, _: B, _: C, _: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe fn default_func(_: A, _: B, _: C, _: D, _: E, _: F) -> R

impl<R, A, B, C, D, E, F> Function<unsafe extern fn(_: A, _: B, _: C, _: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, a: A, b: B, c: C, d: D, e: E, f: F) -> R

fn get(&self) -> unsafe extern fn(_: A, _: B, _: C, _: D, _: E, _: F) -> R

fn set(&mut self, f: unsafe extern fn(_: A, _: B, _: C, _: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern fn default_func(_: A, _: B, _: C, _: D, _: E, _: F) -> R

impl<R, A, B, C, D, E, F> Function<unsafe extern "system" fn(_: A, _: B, _: C, _: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, a: A, b: B, c: C, d: D, e: E, f: F) -> R

fn get(&self)
-> unsafe extern "system" fn(_: A, _: B, _: C, _: D, _: E, _: F) -> R

fn set(&mut self,
f: unsafe extern "system" fn(_: A, _: B, _: C, _: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "system" fn default_func(_: A, _: B, _: C, _: D, _: E, _: F) -> R

impl<R, A, B, C, D, E, F> Function<unsafe extern "fastcall" fn(_: A, _: B, _: C, _: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, a: A, b: B, c: C, d: D, e: E, f: F) -> R

fn get(&self)
-> unsafe extern "fastcall" fn(_: A, _: B, _: C, _: D, _: E, _: F) -> R

fn set(&mut self,
f: unsafe extern "fastcall" fn(_: A, _: B, _: C, _: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "fastcall" fn default_func(_: A,
_: B,
_: C,
_: D,
_: E,
_: F)
-> R

impl<R, B, C, D, E, F> Function<unsafe fn(_: B, _: C, _: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, b: B, c: C, d: D, e: E, f: F) -> R

fn get(&self) -> unsafe fn(_: B, _: C, _: D, _: E, _: F) -> R

fn set(&mut self, f: unsafe fn(_: B, _: C, _: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe fn default_func(_: B, _: C, _: D, _: E, _: F) -> R

impl<R, B, C, D, E, F> Function<unsafe extern fn(_: B, _: C, _: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, b: B, c: C, d: D, e: E, f: F) -> R

fn get(&self) -> unsafe extern fn(_: B, _: C, _: D, _: E, _: F) -> R

fn set(&mut self, f: unsafe extern fn(_: B, _: C, _: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern fn default_func(_: B, _: C, _: D, _: E, _: F) -> R

impl<R, B, C, D, E, F> Function<unsafe extern "system" fn(_: B, _: C, _: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, b: B, c: C, d: D, e: E, f: F) -> R

fn get(&self)
-> unsafe extern "system" fn(_: B, _: C, _: D, _: E, _: F) -> R

fn set(&mut self,
f: unsafe extern "system" fn(_: B, _: C, _: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "system" fn default_func(_: B, _: C, _: D, _: E, _: F) -> R

impl<R, B, C, D, E, F> Function<unsafe extern "fastcall" fn(_: B, _: C, _: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, b: B, c: C, d: D, e: E, f: F) -> R

fn get(&self)
-> unsafe extern "fastcall" fn(_: B, _: C, _: D, _: E, _: F) -> R

fn set(&mut self,
f: unsafe extern "fastcall" fn(_: B, _: C, _: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "fastcall" fn default_func(_: B, _: C, _: D, _: E, _: F) -> R

impl<R, C, D, E, F> Function<unsafe fn(_: C, _: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, c: C, d: D, e: E, f: F) -> R

fn get(&self) -> unsafe fn(_: C, _: D, _: E, _: F) -> R

fn set(&mut self, f: unsafe fn(_: C, _: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe fn default_func(_: C, _: D, _: E, _: F) -> R

impl<R, C, D, E, F> Function<unsafe extern fn(_: C, _: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, c: C, d: D, e: E, f: F) -> R

fn get(&self) -> unsafe extern fn(_: C, _: D, _: E, _: F) -> R

fn set(&mut self, f: unsafe extern fn(_: C, _: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern fn default_func(_: C, _: D, _: E, _: F) -> R

impl<R, C, D, E, F> Function<unsafe extern "system" fn(_: C, _: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, c: C, d: D, e: E, f: F) -> R

fn get(&self)
-> unsafe extern "system" fn(_: C, _: D, _: E, _: F) -> R

fn set(&mut self,
f: unsafe extern "system" fn(_: C, _: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "system" fn default_func(_: C, _: D, _: E, _: F) -> R

impl<R, C, D, E, F> Function<unsafe extern "fastcall" fn(_: C, _: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, c: C, d: D, e: E, f: F) -> R

fn get(&self)
-> unsafe extern "fastcall" fn(_: C, _: D, _: E, _: F) -> R

fn set(&mut self,
f: unsafe extern "fastcall" fn(_: C, _: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "fastcall" fn default_func(_: C, _: D, _: E, _: F) -> R

impl<R, D, E, F> Function<unsafe fn(_: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, d: D, e: E, f: F) -> R

fn get(&self) -> unsafe fn(_: D, _: E, _: F) -> R

fn set(&mut self, f: unsafe fn(_: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe fn default_func(_: D, _: E, _: F) -> R

impl<R, D, E, F> Function<unsafe extern fn(_: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, d: D, e: E, f: F) -> R

fn get(&self) -> unsafe extern fn(_: D, _: E, _: F) -> R

fn set(&mut self, f: unsafe extern fn(_: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern fn default_func(_: D, _: E, _: F) -> R

impl<R, D, E, F> Function<unsafe extern "system" fn(_: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, d: D, e: E, f: F) -> R

fn get(&self) -> unsafe extern "system" fn(_: D, _: E, _: F) -> R

fn set(&mut self, f: unsafe extern "system" fn(_: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "system" fn default_func(_: D, _: E, _: F) -> R

impl<R, D, E, F> Function<unsafe extern "fastcall" fn(_: D, _: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, d: D, e: E, f: F) -> R

fn get(&self) -> unsafe extern "fastcall" fn(_: D, _: E, _: F) -> R

fn set(&mut self,
f: unsafe extern "fastcall" fn(_: D, _: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "fastcall" fn default_func(_: D, _: E, _: F) -> R

impl<R, E, F> Function<unsafe fn(_: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, e: E, f: F) -> R

fn get(&self) -> unsafe fn(_: E, _: F) -> R

fn set(&mut self, f: unsafe fn(_: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe fn default_func(_: E, _: F) -> R

impl<R, E, F> Function<unsafe extern fn(_: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, e: E, f: F) -> R

fn get(&self) -> unsafe extern fn(_: E, _: F) -> R

fn set(&mut self, f: unsafe extern fn(_: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern fn default_func(_: E, _: F) -> R

impl<R, E, F> Function<unsafe extern "system" fn(_: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, e: E, f: F) -> R

fn get(&self) -> unsafe extern "system" fn(_: E, _: F) -> R

fn set(&mut self, f: unsafe extern "system" fn(_: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "system" fn default_func(_: E, _: F) -> R

impl<R, E, F> Function<unsafe extern "fastcall" fn(_: E, _: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, e: E, f: F) -> R

fn get(&self) -> unsafe extern "fastcall" fn(_: E, _: F) -> R

fn set(&mut self, f: unsafe extern "fastcall" fn(_: E, _: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "fastcall" fn default_func(_: E, _: F) -> R

impl<R, F> Function<unsafe fn(_: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, f: F) -> R

fn get(&self) -> unsafe fn(_: F) -> R

fn set(&mut self, f: unsafe fn(_: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe fn default_func(_: F) -> R

impl<R, F> Function<unsafe extern fn(_: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, f: F) -> R

fn get(&self) -> unsafe extern fn(_: F) -> R

fn set(&mut self, f: unsafe extern fn(_: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern fn default_func(_: F) -> R

impl<R, F> Function<unsafe extern "system" fn(_: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, f: F) -> R

fn get(&self) -> unsafe extern "system" fn(_: F) -> R

fn set(&mut self, f: unsafe extern "system" fn(_: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "system" fn default_func(_: F) -> R

impl<R, F> Function<unsafe extern "fastcall" fn(_: F) -> R>

fn is_default(&self) -> bool

unsafe fn call(&self, f: F) -> R

fn get(&self) -> unsafe extern "fastcall" fn(_: F) -> R

fn set(&mut self, f: unsafe extern "fastcall" fn(_: F) -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "fastcall" fn default_func(_: F) -> R

impl<R> Function<unsafe fn() -> R>

fn is_default(&self) -> bool

unsafe fn call(&self) -> R

fn get(&self) -> unsafe fn() -> R

fn set(&mut self, f: unsafe fn() -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe fn default_func() -> R

impl<R> Function<unsafe extern fn() -> R>

fn is_default(&self) -> bool

unsafe fn call(&self) -> R

fn get(&self) -> unsafe extern fn() -> R

fn set(&mut self, f: unsafe extern fn() -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern fn default_func() -> R

impl<R> Function<unsafe extern "system" fn() -> R>

fn is_default(&self) -> bool

unsafe fn call(&self) -> R

fn get(&self) -> unsafe extern "system" fn() -> R

fn set(&mut self, f: unsafe extern "system" fn() -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "system" fn default_func() -> R

impl<R> Function<unsafe extern "fastcall" fn() -> R>

fn is_default(&self) -> bool

unsafe fn call(&self) -> R

fn get(&self) -> unsafe extern "fastcall" fn() -> R

fn set(&mut self, f: unsafe extern "fastcall" fn() -> R)

unsafe fn set_from_raw(&mut self, f: *const c_void)

unsafe extern "fastcall" fn default_func() -> R