Function hl_capture::hooks::hw::RunListenServer

#[export_name = "_Z15RunListenServerPvPcS0_S0_PFP14IBaseInterfacePKcPiES7_"]

pub unsafe extern fn RunListenServer(instance: *mut c_void,
                                     basedir: *mut c_char,
                                     cmdline: *mut c_char,
                                     postRestartCmdLineArgs: *mut c_char,
                                     launcherFactory: *mut c_void,
                                     filesystemFactory: *mut c_void)
                                     -> c_int

The "main" function of hw.so, called inside CEngineAPI::Run().

The game runs within this function and shortly after it exits hw.so is unloaded. Note: _restart also causes this function to exit, in this case the launcher unloads and reloads hw.so and this function is called again as if it was a fresh start.